// src/models/tmdb/person/PersonDetails.ts
export interface PersonDetails {
  id: number;
  adult: boolean;
  biography: string;
  birthday: string;
  deathday: null | string;
  gender: number;
  homepage: null | string;
  imdb_id: number;
  known_for_department: string;
  name: string;
  place_of_birth: string;
  profile_path: string;
  popularity: number;
}